Testes integrados do endpoint de criação e de get por genero por id

+++ admin-do-catalogo/infrastructure/src/main/java/org/jonasribeiro/admin/catalogo/infraestructure/genre/GenreMySQLGateway.java
@@ -32,7 +32,9 @@ public class GenreMySQLGateway implements GenreGateway {
 
     @Override
     public void deleteById(GenreID anId) {
-        this.genreRepository.deleteById(anId.getValue());
+        if(this.genreRepository.existsById(anId.getValue())) {
+            this.genreRepository.deleteById(anId.getValue());
+        }
     }
 
     @Override
